requirements.txt
huggingface-hub>=0.19.3,<1.0
transformers>=4.35.0
torch>=2.0.0
gradio>=4.0.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib